buscar medico

+++ src/main/java/alto/grupo/entidades/Medico.java
@@ -44,7 +44,7 @@ public class Medico {
     @ElementCollection
     @CollectionTable(name = "medico_centros_medicos", joinColumns = @JoinColumn(name = "medico_matricula"))
     @Column(name = "centro_medico")
-    private List<Integer> centrosMedicos;  //
+    private List<Long> centrosMedicos;  //
 
     public Medico() {
     }
@@ -103,7 +103,7 @@ public class Medico {
         return especialidad3;
     }
 
-    public List<Integer> getCentrosMedicos() {
+    public List<Long> getCentrosMedicos() {
         return centrosMedicos;
     }
 
@@ -159,7 +159,7 @@ public class Medico {
         this.especialidad3 = especialidad3;
     }
 
-    public void setCentrosMedicos(List<Integer> centrosMedicos) {
+    public void setCentrosMedicos(List<Long> centrosMedicos) {
         this.centrosMedicos = centrosMedicos;
     }
